Change the database name to `amqonline`
In 1.0.x, the postgresql database name was `amqonline`. I believe this was determined by the `database-name` key in the `postgresql` secret created by the keycloak-controller.

However, the secret is now created elsewhere, and sets a value based on the `database` key here, which is hardcoded as `postgresql` now.
This causes 2 problems for us that we have to try work around.

- existing installations being upgraded to 1.0 1.1.1 need to patch the AuthenticationService after the upgrade to reset the database name to amqonline. We don't want to lose any data here, so we have to ensure it's using the same database.
- new installations of 1.1.1 need to also patch the AuthenticationService to set the database name to `amqonline` to avoid different clusters having different setups

Ideally the database name stays the same as it was before (`amqonline`).

cc @k-wall for discussion on this.